+++ adapters/coroutines/src/main/kotlin/xyz/urbanmatrix/mavlink/adapters/coroutines/CoroutinesMavConnection.kt
@@ -18,25 +18,25 @@ public interface CoroutinesMavConnection {
 
     @Throws(IOException::class)
     public suspend fun <T : MavMessage<T>> sendV1(
-        systemId: Int,
-        componentId: Int,
+        systemId: UByte,
+        componentId: UByte,
         payload: T
     )
 
     @Throws(IOException::class)
     public suspend fun <T : MavMessage<T>> sendUnsignedV2(
-        systemId: Int,
-        componentId: Int,
+        systemId: UByte,
+        componentId: UByte,
         payload: T
     )
 
     @Throws(IOException::class)
     public suspend fun <T : MavMessage<T>> sendSignedV2(
-        systemId: Int,
-        componentId: Int,
+        systemId: UByte,
+        componentId: UByte,
         payload: T,
-        linkId: Int,
-        timestamp: Long,
+        linkId: UByte,
+        timestamp: UInt,
         secretKey: ByteArray
     )
 }
